feat: admin용 API로 변경 (테마, 스케줄 생성/삭제)
